requirements.txt
pygame-ce
pytest
ruff
black
mypy
